+++ server/test_data/test_issues.jsx
@@ -217,6 +217,7 @@ TestIssues.generate = function() {
 	    createdAt:randomDate(startDate, new Date()),
 	    priority:newStatus=='Closed'?'Closed':getRandom(['Scheduled','Standard','Urgent','Critical']),
 	    thumb:["img/issue-1.jpg","img/issue-2.jpg","img/issue-3.jpg"],
+	    closeDetails:{},
 	    contact:{
 	      name:"John Smith",
 	      email:"johnny@flart.flart",